# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/analytics/dashboard": {
      "get": {
        "operationId": "AnalyticsController_getDashboard_v1",
        "parameters": [
          {
            "name": "x-company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/analytics/search": {
      "get": {
        "operationId": "AnalyticsController_globalSearch_v1",
        "parameters": [
          {
            "name": "x-company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/v1/analytics/report": {
      "get": {
        "operationId": "AnalyticsController_getReport_v1",
        "parameters": [
          {
            "name": "x-company-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    }
  },
  "info": {
    "title": "Smatal ERP API",
    "description": "Enterprise ERP Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}